# [5.1] Fix pagination for group by and having statements