There are two problems that this PR is trying to address:

1.) length-aware pagination with "group by" statements is currently slow since counting of the results is done in PHP instead of in the query
2.) and it's currently not possible to paginate queries containing having statements

In these two cases, we can use a sub query to get the total number of rows:
```
select count(*) as aggregate from (<original query>) AS aggregate_table
```

By doing this, we can now paginate any query and it's all done in SQL so it should be a lot faster. If no "groups" or "havings" are defined, then we still use the normal pagination without a subquery. And since we don't need the "hack" for group by statements any more, I was able to clean up some related code as well and defer to the built-in count function in both cases.

This was tested in SQLite and MySQL. Any feedback about other supported grammars is welcome.