@GrahamCampbell 2015-10-16
https://styleci.io/analyses/zEKYAX