@taylorotwell 2015-10-16
Problem with this. Say you have a 1,000,000 record user table. Instead of just running a query like `select count(*) from users`... now you are running `select count(*) from (select * from users) as ...`

You are selecting all 1,000,000 rows and loading them up, then counting them. That's not going to be performant.