@acasar 2015-10-16
@taylorotwell  Well, until now we were doing the same, just loading them all in PHP and counting them there. I think this is still much better.

And once again - it doesn't change anything if you don't have "group by" or "having" statements - in these cases a normal ``select count(*) from users...`` is executed. This PR just takes care of some edge cases which didn't work before (having statements) and improves performance for "group by" statements.