@taylorotwell 2015-10-16
No, that's not what we're doing right now. Here are the two queries generated from `User::paginate()` out of the box:

```
"select count(*) as aggregate from `users`"
"select * from `users` limit 15 offset 0"
```