@taylorotwell 2015-10-16
You can see that we're just doing a count, we're not actually pulling back the entire users table.