@acasar 2015-10-16
And that still remains the same. Try adding a group by and see what happens now vs. after.