@acasar 2015-10-16
@taylorotwell I made some benchmarks on a table with 100,000 rows and I ran the following query 100 times:

```
EloquentTestUser::groupBy('email')->getQuery()->getCountForPagination()
```

The results show that with this PR it's about 20% faster.

Without PR: 12.6s, 12.9s
With PR: 10,7s, 10,6s