@tylercrompton 2015-11-02
Which database did you use? Did you try others. The results are possibly very implementation-specific.